requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.3
django-admin-thumbnails==0.2.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
